backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
motor==4.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.28.47
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.0
openai==1.12.0
boto3==1.34.0
email-validator==1.3.1
